clean up code, drop old synapse implementation
