+++ composer.json
@@ -40,14 +40,7 @@
         "tests-storage": "phpunit --colors=always --testsuite storage",
         "tests-snowflake-abs": "SUITE=tests-snowflake STORAGE_TYPE=ABS BUILD_PREFIX=$BUILD_PREFIX phpunit --colors=always --testsuite snowflake",
         "tests-snowflake-s3": "SUITE=tests-snowflake-s3  STORAGE_TYPE=S3 BUILD_PREFIX=$BUILD_PREFIX phpunit --colors=always --testsuite snowflake-s3",
-        "tests-synapse": "SUITE=tests-synapse CREDENTIALS_IMPORT_TYPE=SAS CREDENTIALS_EXPORT_TYPE=MASTER_KEY TEMP_TABLE_TYPE=HEAP DEDUP_TYPE=TMP_TABLE phpunit --colors=always --testsuite synapse",
-        "tests-synapse-heap4000temp-optimized": "SUITE=tests-synapse-heap4000temp-optimized CREDENTIALS_IMPORT_TYPE=SAS CREDENTIALS_EXPORT_TYPE=MASTER_KEY TEMP_TABLE_TYPE=HEAP4000 DEDUP_TYPE=CTAS phpunit --colors=always --testsuite tests-synapse-heap4000temp-optimized",
-        "tests-synapse-heap4000temp-optimized-hash": "SUITE=tests-synapse-heap4000temp-optimized-hash CREDENTIALS_IMPORT_TYPE=SAS CREDENTIALS_EXPORT_TYPE=MASTER_KEY TEMP_TABLE_TYPE=HEAP4000 DEDUP_TYPE=CTAS TABLE_DISTRIBUTION=HASH phpunit --colors=always --testsuite tests-synapse-heap4000temp-optimized",
-        "tests-synapse-next": "SUITE=tests-synapse-next CREDENTIALS_IMPORT_TYPE=SAS CREDENTIALS_EXPORT_TYPE=MASTER_KEY TEMP_TABLE_TYPE=HEAP4000 DEDUP_TYPE=CTAS TABLE_DISTRIBUTION=HASH phpunit --colors=always --testsuite tests-synapse-next",
-        "tests-synapse-columnstoretemp": "SUITE=tests-synapse-columnstoretemp CREDENTIALS_IMPORT_TYPE=SAS CREDENTIALS_EXPORT_TYPE=MASTER_KEY TEMP_TABLE_TYPE=COLUMNSTORE DEDUP_TYPE=TMP_TABLE phpunit --colors=always --testsuite tests-synapse-columnstoretemp",
-        "tests-synapse-heap4000temp": "SUITE=tests-synapse-heap4000temp CREDENTIALS_IMPORT_TYPE=SAS CREDENTIALS_EXPORT_TYPE=MASTER_KEY TEMP_TABLE_TYPE=HEAP4000 DEDUP_TYPE=TMP_TABLE phpunit --colors=always --testsuite tests-synapse-heap4000temp",
-        "tests-synapse-heap4000temp-hash": "SUITE=tests-synapse-heap4000temp-hash CREDENTIALS_IMPORT_TYPE=SAS CREDENTIALS_EXPORT_TYPE=MASTER_KEY TEMP_TABLE_TYPE=HEAP4000 DEDUP_TYPE=TMP_TABLE TABLE_DISTRIBUTION=HASH phpunit --colors=always --testsuite tests-synapse-heap4000temp",
-        "tests-synapse-clusterdindextemp": "SUITE=tests-synapse-clusterdindextemp CREDENTIALS_IMPORT_TYPE=SAS CREDENTIALS_EXPORT_TYPE=MASTER_KEY TEMP_TABLE_TYPE=CLUSTERED_INDEX DEDUP_TYPE=TMP_TABLE phpunit --colors=always --testsuite tests-synapse-clusterdindextemp",
+        "tests-synapse": "SUITE=tests-synapse CREDENTIALS_IMPORT_TYPE=SAS CREDENTIALS_EXPORT_TYPE=MASTER_KEY TEMP_TABLE_TYPE=HEAP4000 DEDUP_TYPE=CTAS TABLE_DISTRIBUTION=HASH phpunit --colors=always --testsuite tests-synapse-next",
         "tests-synapse-mi": "SUITE=tests-synapse-mi CREDENTIALS_IMPORT_TYPE=MANAGED_IDENTITY CREDENTIALS_EXPORT_TYPE=MANAGED_IDENTITY TEMP_TABLE_TYPE=HEAP DEDUP_TYPE=TMP_TABLE phpunit --colors=always --testsuite synapse-mi",
         "tests-exasol": "SUITE=tests-exasol STORAGE_TYPE=S3 phpunit --colors=always --testsuite exasol",
         "tests-teradata-tpt": "SUITE=tests-teradata-tpt TERADATA_CSV_ADAPTER=TPT STORAGE_TYPE=S3 phpunit --colors=always --testsuite tests-teradata",
@@ -56,10 +49,6 @@
             "@tests-snowflake-s3",
             "@tests-synapse",
             "@tests-synapse-mi",
-            "@tests-synapse-columnstoretemp",
-            "@tests-synapse-clusterdindextemp",
-            "@tests-synapse-heap4000temp",
-            "@tests-synapse-heap4000temp-optimized",
             "@tests-exasol",
             "@tests-teradata-tpt"
         ],
