+++ tests/functional/Synapse/IncrementalImportNoTypesTest.php
@@ -2,9 +2,8 @@
 
 declare(strict_types=1);
 
-namespace Tests\Keboola\Db\ImportExportFunctional\SynapseNext;
+namespace Tests\Keboola\Db\ImportExportFunctional\Synapse;
 
-use Keboola\Csv\CsvFile;
 use Keboola\Db\ImportExport\Backend\Synapse\ToFinalTable\FullImporter;
 use Keboola\Db\ImportExport\Backend\Synapse\ToFinalTable\IncrementalImporter;
 use Keboola\Db\ImportExport\Backend\Synapse\ToFinalTable\SqlBuilder;
