+++ tests/functional/Synapse/SynapseBaseTestCase.php
@@ -7,11 +7,17 @@ namespace Tests\Keboola\Db\ImportExportFunctional\Synapse;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Platforms\SQLServer2012Platform;
-use Keboola\Db\ImportExport\Backend\Synapse\SqlCommandBuilder;
+use Keboola\Datatype\Definition\Synapse;
+use Keboola\Db\ImportExport\Backend\Synapse\ToFinalTable\SqlBuilder;
 use Keboola\Db\ImportExport\ImportOptions;
 use Keboola\Db\ImportExport\Storage\SourceInterface;
-use Keboola\Db\ImportExport\Storage\Synapse\Table;
 use Keboola\Db\ImportExport\Backend\Synapse\SynapseImportOptions;
+use Keboola\Db\ImportExport\Storage\Synapse\Table;
+use Keboola\TableBackendUtils\Column\ColumnCollection;
+use Keboola\TableBackendUtils\Column\SynapseColumn;
+use Keboola\TableBackendUtils\Table\Synapse\TableDistributionDefinition;
+use Keboola\TableBackendUtils\Table\Synapse\TableIndexDefinition;
+use Keboola\TableBackendUtils\Table\SynapseTableDefinition;
 use Keboola\TableBackendUtils\Table\SynapseTableReflection;
 use Tests\Keboola\Db\ImportExportFunctional\ImportExportBaseTest;
 
@@ -34,7 +40,7 @@ class SynapseBaseTestCase extends ImportExportBaseTest
     /** @var Connection */
     protected $connection;
 
-    /** @var SqlCommandBuilder */
+    /** @var SqlBuilder */
     protected $qb;
 
     /** @var SQLServer2012Platform|AbstractPlatform */
@@ -77,15 +83,15 @@ EOT
     public function getSourceSchemaName(): string
     {
         return self::SYNAPSE_SOURCE_SCHEMA_NAME
-        . '-'
-        . getenv('SUITE');
+            . '-'
+            . getenv('SUITE');
     }
 
     public function getDestinationSchemaName(): string
     {
         return self::SYNAPSE_DEST_SCHEMA_NAME
-        . '-'
-        . getenv('SUITE');
+            . '-'
+            . getenv('SUITE');
     }
 
     protected function initTables(array $tables): void
@@ -266,7 +272,7 @@ EOT
     {
         parent::setUp();
         $this->connection = $this->getSynapseConnection();
-        $this->qb = new SqlCommandBuilder($this->connection);
+        $this->qb = new SqlBuilder();
         $this->platform = $this->connection->getDatabasePlatform();
     }
 
@@ -295,6 +301,7 @@ EOT
 
     /**
      * @param int|string $sortKey
+     * @param array<mixed> $expected
      */
     protected function assertTableEqualsExpected(
         SourceInterface $source,
@@ -382,5 +389,121 @@ EOT
             // @phpstan-ignore-next-line
             getenv('DEDUP_TYPE')
         );
+    }
+
+    /**
+     * @param string[] $columnsNames
+     */
+    public function getColumnsWithoutTypes(array $columnsNames): ColumnCollection
+    {
+        $columns = array_map(function ($colName) {
+            return new SynapseColumn(
+                $colName,
+                new Synapse(
+                    'NVARCHAR',
+                    ['length' => 4000]
+                )
+            );
+        }, $columnsNames);
+        return new ColumnCollection($columns);
+    }
+
+    /**
+     * @param string[] $columns
+     * @param string[] $pks
+     */
+    public function getGenericTableDefinition(
+        string $schemaName,
+        string $tableName,
+        array $columns,
+        array $pks = []
+    ): SynapseTableDefinition {
+        return new SynapseTableDefinition(
+            $schemaName,
+            $tableName,
+            false,
+            $this->getColumnsWithoutTypes($columns),
+            $pks,
+            new TableDistributionDefinition(TableDistributionDefinition::TABLE_DISTRIBUTION_ROUND_ROBIN, []),
+            new TableIndexDefinition(TableIndexDefinition::TABLE_INDEX_TYPE_CLUSTERED_INDEX, [])
+        );
+    }
+
+    /**
+     * @param int|string $sortKey
+     * @param array<mixed> $expected
+     * @param string|int $sortKey
+     */
+    protected function assertSynapseTableEqualsExpected(
+        SourceInterface $source,
+        SynapseTableDefinition $destination,
+        SynapseImportOptions $options,
+        array $expected,
+        $sortKey,
+        string $message = 'Imported tables are not the same as expected'
+    ): void {
+        $tableColumns = (new SynapseTableReflection(
+            $this->connection,
+            $destination->getSchemaName(),
+            $destination->getTableName()
+        ))->getColumnsNames();
+
+        if ($options->useTimestamp()) {
+            self::assertContains('_timestamp', $tableColumns);
+        } else {
+            self::assertNotContains('_timestamp', $tableColumns);
+        }
+
+        if (!in_array('_timestamp', $source->getColumnsNames())) {
+            $tableColumns = array_filter($tableColumns, function ($column) {
+                return $column !== '_timestamp';
+            });
+        }
+
+        $tableColumns = array_map(function ($column) {
+            return sprintf('[%s]', $column);
+        }, $tableColumns);
+
+        $sql = sprintf(
+            'SELECT %s FROM [%s].[%s]',
+            implode(', ', $tableColumns),
+            $destination->getSchemaName(),
+            $destination->getTableName()
+        );
+
+        $queryResult = array_map(function ($row) {
+            return array_map(function ($column) {
+                return $column;
+            }, array_values($row));
+        }, $this->connection->fetchAllAssociative($sql));
+
+        $this->assertArrayEqualsSorted(
+            $expected,
+            $queryResult,
+            $sortKey,
+            $message
+        );
+    }
+
+    protected function assertSynapseTableExpectedRowCount(
+        SynapseTableDefinition $destination,
+        SynapseImportOptions $options,
+        int $expectedCount,
+        string $message = 'Imported tables don\'t have expected number of rows'
+    ): void {
+        $destRef = new SynapseTableReflection(
+            $this->connection,
+            $destination->getSchemaName(),
+            $destination->getTableName()
+        );
+        $tableColumns = $destRef->getColumnsNames();
+
+        if ($options->useTimestamp()) {
+            self::assertContains('_timestamp', $tableColumns);
+        } else {
+            self::assertNotContains('_timestamp', $tableColumns);
+        }
+
+        self::assertEquals($expectedCount, $destRef->getRowsCount(), $message);
     }
 }
