+++ tests/unit/Storage/Synapse/TableTest.php
@@ -4,14 +4,7 @@ declare(strict_types=1);
 
 namespace Tests\Keboola\Db\ImportExportUnit\Storage\Synapse;
 
-use Keboola\Db\Import\Result;
-use Keboola\Db\ImportExport\Backend\BackendImportAdapterInterface;
-use Keboola\Db\ImportExport\Backend\ImporterInterface;
-use Keboola\Db\ImportExport\Backend\Synapse\Importer as SynapseImporter;
-use Keboola\Db\ImportExport\ImportOptions;
-use Keboola\Db\ImportExport\Storage\NoBackendAdapterException;
 use Keboola\Db\ImportExport\Storage;
-use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
 class TableTest extends TestCase
